added user overview component
added to test page as well
