+++ src/lib/components/overview/index.ts
@@ -1,2 +1,3 @@
 export { default as GroupOverview } from './GroupOverview.svelte';
 export { default as ThreadOverview } from './ThreadOverview.svelte';
+export { default as UserOverview } from './UserOverview.svelte';
