+++ src/routes/(app)/users/+page.svelte
@@ -1,30 +1,20 @@
 <script lang="ts">
-  import { enhance } from '$app/forms';
   import { goto } from '$app/navigation';
   import { page } from '$app/state';
   import { Button } from '$lib/components/button';
   import { Input } from '$lib/components/input';
   import { Scroller } from '$lib/components/layout';
   import { Icon, ReturnHeader } from '$lib/components/misc';
-  import { Modal, ModalBody, ModalFooter, ModalHeader } from '$lib/components/modal';
-  import { getToaster } from '$lib/components/toast';
-  import { formResultToast } from '$lib/utils/helpers';
+  import { UserOverview } from '$lib/components/overview';
   import type { PageProps } from './$types';
 
   let { data }: PageProps = $props();
 
-  const toaster = getToaster();
-  let showModal = $state<'delete' | null>(null);
-  let username = $state<string>('');
   let pageIndex = 1;
   let disableScroller = $state<boolean>(data.endOfList);
   let users = $derived(data.users);
   let searchString = $derived<string>(decodeURIComponent(page.url.searchParams.get('q') ?? ''));
   let searched = $derived<boolean>(page.url.searchParams.has('q'));
-
-  function hideModal() {
-    showModal = null;
-  }
 </script>
 
 <svelte:head>
@@ -71,33 +61,7 @@
     <p class="p-4 text-center text-xl font-semibold">No result found.</p>
   {:else}
     {#each users as user}
-      <a class="flex items-center gap-4 box box-hover" href="/user/{user.username}">
-        <img
-          class="profile-picture-md"
-          src={user.profilePicture ?? '/images/default-user-profile-icon.jpg'}
-          alt="profile"
-        />
-        <div class="flex flex-col gap-2">
-          <p class="text-gray-700">#{user.id}</p>
-          <div class="flex items-baseline gap-2">
-            <span class="text-xl font-semibold">{user.displayName}</span>
-            <span class="text-gray-500">@{user.username}</span>
-          </div>
-          <p>Role: {user.role}</p>
-        </div>
-
-        {#if user.role !== 'admin'}
-          <Button
-            class="ml-auto p-2"
-            onclick={() => {
-              showModal = 'delete';
-              username = user.username;
-            }}
-          >
-            <Icon type="delete" class="text-red-500" />
-          </Button>
-        {/if}
-      </a>
+      <UserOverview {user} />
     {/each}
   {/if}
 
@@ -118,28 +82,3 @@
     }}
   />
 </div>
-
-<Modal show={showModal === 'delete'} backdropCallback={hideModal} center>
-  <ModalHeader>Delete account</ModalHeader>
-  <ModalBody>
-    Are you sure you want to delete this account? This action is irreversible and all data of this
-    user can not be recovered. This user will no longer be able to login.
-  </ModalBody>
-  <ModalFooter>
-    <form
-      class="w-full"
-      action="?/delete-account"
-      method="post"
-      use:enhance={() => {
-        return async ({ result, update }) => {
-          await formResultToast(result, toaster);
-          await update();
-        };
-      }}
-    >
-      <Button class="w-full" type="danger" onclick={hideModal}>Delete</Button>
-      <input type="text" name="username" value={username} hidden readonly />
-    </form>
-    <Button class="w-full" type="base" onclick={hideModal}>Cancel</Button>
-  </ModalFooter>
-</Modal>
